Blender 5.0 - accomodate new mathutils buffer protocol type
Long story short - since 5.0 `np.array(Vector())` is now producing `np.float32` instead of `np.float64`. So we have to provide `dtype` explicitly to support both <5.0 and >= 5.0.

See https://projects.blender.org/blender/blender/issues/149283

+++ src/bonsai/bonsai/tool/geometry.py
@@ -1146,8 +1146,9 @@ class Geometry(bonsai.core.tool.Geometry):
     def record_object_position(cls, obj: bpy.types.Object) -> None:
         # These are recorded separately because they have different numerical tolerances
         props = tool.Blender.get_object_bim_props(obj)
-        props.location_checksum = repr(np.array(obj.matrix_world.translation).tobytes())
-        props.rotation_checksum = repr(np.array(obj.matrix_world.to_3x3()).tobytes())
+        # Explicit dtype for Blender <5.0 compatibility.
+        props.location_checksum = repr(np.array(obj.matrix_world.translation, dtype=np.float32).tobytes())
+        props.rotation_checksum = repr(np.array(obj.matrix_world.to_3x3(), dtype=np.float32).tobytes())
 
     @classmethod
     def remove_connection(cls, connection: ifcopenshell.entity_instance) -> None:
